physical: added xsd schema to france database adaptation: added xsd reference on saving file

+++ src/applications/adaptation_app/ADN_Equipement_Postures_GUI.cpp
@@ -69,7 +69,7 @@ ADN_Equipement_Postures_GUI::ADN_Equipement_Postures_GUI( const QString& strColC
     setSorting( true );
     setShowGrid( false );
     setLeftMargin( 0 );
-    setMinimumHeight( ( eNbrUnitPosture + 1.2 ) * 20 );
+    setMinimumHeight( int( ( eNbrUnitPosture + 1.2 ) * 20 ) );
 
     // Setup the columns and headers
     setNumCols( 2 );
