+++ src/applications/adaptation_app/ADN_Sensors_Postures_GUI.cpp
@@ -69,7 +69,7 @@ ADN_Sensors_Postures_GUI::ADN_Sensors_Postures_GUI( const QString& strColCaption
     setSorting( true );
     setShowGrid( false );
     setLeftMargin( 0 );
-    setMinimumHeight( ( eNbrUnitPosture + 1.2 ) * 20 );
+    setMinimumHeight( int( ( eNbrUnitPosture + 1.2 ) * 20 ) );
 
     // Setup the columns and headers
     setNumCols( 2 );
